@@ -14,32 +14,33 @@
   "dependencies": {
     "@fortawesome/fontawesome": "^1.1.8",
     "@fortawesome/fontawesome-free-solid": "^5.0.13",
-    "@fortawesome/fontawesome-svg-core": "^6.6.0",
-    "@fortawesome/free-solid-svg-icons": "^6.7.1",
-    "@fortawesome/vue-fontawesome": "^3.0.3",
-    "@kolirt/vue-modal": "^1.0.3",
-    "@popperjs/core": "^2.11.6",
-    "@vue/runtime-dom": "^3.4.35",
-    "axios": "^1.7.2",
-    "bootstrap": "^5.2.3",
+    "@fortawesome/fontawesome-svg-core": "^6.7.2",
+    "@fortawesome/free-solid-svg-icons": "^6.7.2",
+    "@fortawesome/vue-fontawesome": "^3.0.8",
+    "@kolirt/vue-modal": "^1.1.2",
+    "@popperjs/core": "^2.11.8",
+    "@stripe/stripe-js": "^6.1.0",
+    "@vue/runtime-dom": "^3.5.13",
+    "axios": "^1.8.4",
+    "bootstrap": "^5.3.3",
     "bootstrap-vue-3": "^0.3.12",
-    "chart.js": "^4.3.0",
+    "chart.js": "^4.4.8",
     "cookieconsent": "^3.1.1",
-    "dotenv": "^16.3.1",
-    "dotenv-webpack": "^8.0.1",
+    "dotenv": "^16.4.7",
     "html-loader": "^5.1.0",
-    "jquery": "^3.7.0",
+    "jquery": "^3.7.1",
     "lodash": "^4.17.21",
-    "moment": "^2.29.4",
-    "moment-timezone": "^0.5.43",
-    "portfinder": "^1.0.32",
-    "swiper": "^11.2.1",
-    "vue": "^3.4.35",
+    "moment": "^2.30.1",
+    "moment-timezone": "^0.5.48",
+    "portfinder": "^1.0.35",
+    "save-dev": "^0.0.1-security",
+    "swiper": "^11.2.6",
+    "vue": "^3.5.13",
     "vue-axios": "^3.5.2",
-    "vue-chartjs": "^5.2.0",
-    "vue-i18n": "^9.2.2",
-    "vue-multiselect": "^3.0.0-beta.3",
-    "vue-router": "^4.4.1",
+    "vue-chartjs": "^5.3.2",
+    "vue-i18n": "^11.1.2",
+    "vue-multiselect": "^3.2.0",
+    "vue-router": "^4.5.0",
     "vue-slider-component": "^4.1.0-beta.7",
     "vue3-google-login": "^2.0.33",
     "vue3-markdown-it": "^1.0.10",
@@ -47,33 +48,34 @@
     "webpack-bundle-analyzer": "^4.10.2"
   },
   "devDependencies": {
-    "@babel/core": "^7.25.2",
-    "@babel/plugin-transform-runtime": "^7.24.7",
-    "@babel/preset-env": "^7.25.3",
-    "@vue/babel-plugin-jsx": "^1.2.2",
-    "@vue/compiler-sfc": "^3.4.35",
-    "@vue/test-utils": "^2.3.2",
-    "autoprefixer": "^10.4.14",
-    "babel-loader": "^9.1.2",
+    "@babel/core": "^7.26.10",
+    "@babel/plugin-transform-runtime": "^7.26.10",
+    "@babel/preset-env": "^7.26.9",
+    "@vue/babel-plugin-jsx": "^1.4.0",
+    "@vue/compiler-sfc": "^3.5.13",
+    "@vue/test-utils": "^2.4.6",
+    "autoprefixer": "^10.4.21",
+    "babel-loader": "^10.0.0",
     "clean-webpack-plugin": "^4.0.0",
-    "copy-webpack-plugin": "^11.0.0",
-    "css-loader": "^6.11.0",
-    "css-minimizer-webpack-plugin": "^7.0.0",
+    "copy-webpack-plugin": "^13.0.0",
+    "css-loader": "^7.1.2",
+    "css-minimizer-webpack-plugin": "^7.0.2",
+    "dotenv-webpack": "^8.1.0",
     "file-loader": "^6.2.0",
-    "html-webpack-plugin": "^5.5.3",
-    "jest": "^29.5.0",
-    "mini-css-extract-plugin": "^2.9.0",
-    "postcss-loader": "^7.3.4",
-    "sass": "^1.77.8",
-    "sass-loader": "^13.3.3",
-    "style-loader": "^3.3.4",
+    "html-webpack-plugin": "^5.6.3",
+    "jest": "^29.7.0",
+    "mini-css-extract-plugin": "*",
+    "postcss-loader": "^8.1.1",
+    "sass": "^1.86.1",
+    "sass-loader": "*",
+    "style-loader": "*",
     "url-loader": "^4.1.1",
-    "vue-loader": "^16.8.3",
+    "vue-loader": "^17.4.2",
     "vue-style-loader": "^4.1.3",
-    "vue-template-compiler": "^0.1.0",
-    "webpack": "^5.93.0",
-    "webpack-cli": "^4.10.0",
-    "webpack-dev-server": "^4.15.2",
-    "webpack-merge": "^5.9.0"
+    "vue-template-compiler": "^2.7.16",
+    "webpack": "^5.98.0",
+    "webpack-cli": "^6.0.1",
+    "webpack-dev-server": "^5.2.1",
+    "webpack-merge": "^6.0.1"
   }
 }
