@@ -4,7 +4,6 @@
     <img class="position-fixed h-100 w-100 z-n1" src="./assets/background.png">
     <nav id="mainNav" class="navbar navbar-expand-lg navbar-light text-uppercase p-2 px-3 w-75">
       <router-link class="navbar-brand" to="/">
-        <!-- <img src="./assets/refairme_logo.png" height="50"> -->
         <img src="./assets/refair-me-logo.svg" height="70" class="ml-2">
       </router-link>
       <button class="navbar-toggler navbar-toggler-right rounded" type="button" data-toggle="collapse" data-target="#navbarResponsive" aria-controls="navbarResponsive" aria-expanded="false" aria-label="Toggle navigation">
@@ -28,25 +27,21 @@
             <router-link class="nav-link py-3 px-0 px-lg-3" to="/auth/signin">Signin</router-link>
           </li>
           <li class="nav-item mx-0 mx-lg-1" v-if="isAuthenticated">
-            <span class="nav-link py-3 px-0 px-lg-3" style="cursor: pointer" v-on:click="logout">Logout</span>
+            <span class="nav-link py-3 px-0 px-lg-3" style="cursor: pointer" @click="logout">Logout</span>
           </li>
         </ul>
       </div>
     </nav>
-    <div class="container p-0" v-bind:class="[isPathHome ? 'max-width' : '']">
+    <div class="container p-0" :class="[isPathHome ? 'max-width' : '']">
       <router-view></router-view>
     </div>
   </div>
-
 </template>
 
 <script>
 import store from '@/store/index.js'
 import cookieconsent from 'cookieconsent'
-
 export default {
-  store,
-
   computed: {
     isAuthenticated: vm => vm.$store.getters.isAuthenticated,
     path: vm => vm.$route.path,
@@ -89,9 +84,8 @@ export default {
   }
 }
 </script>
+
 <style lang="scss">
-//@import '../node_modules/bootstrap/scss/bootstrap.scss';
-@import url('https://fonts.googleapis.com/css?family=Source+Sans+Pro');
 @import '@/assets/settings.scss';
 @import url('https://cdn.jsdelivr.net/npm/vue-slider-component@latest/theme/default.css');
 
@@ -125,6 +119,4 @@ nav {
   background-color: white;
   border-radius: .375rem;
 }
-
-
 </style>
